requirements.txt
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
Pillow==9.0.1
h5py==3.7.0
imutils==0.5.3

numpy~=1.23.5
mtcnn~=0.1.1
scipy~=1.9.3
tensorflow~=2.11.0
requests~=2.28.1
scikit-learn~=1.2.0
keras~=2.11.0
Flask~=2.2.3